Script for extracting values for MK tests
Obtains values for McDonald-Kreitman tests. Uses Table S3 from Fraisse et al, 2017 and lists of genes of interest to output 1) a list of genes of interest missing in Table S3 and 2) a table of values for the genes of interest found in Table S3. Again, file names are hard coded.